requirements.txt
requests
pyqt5
keyboard
mecab-python3
PyYAML